services/ingestion/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
neo4j==5.15.0
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
unstructured==0.11.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.54.0
pydantic==2.5.0
tenacity==8.2.3
